+++ bc/recruitment/models.py
@@ -2,6 +2,7 @@ import json
 import secrets
 from urllib.parse import urlsplit, urlunsplit
 
+from django import forms
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.db.models import Count, F
@@ -23,11 +24,11 @@ from wagtail.search import index
 
 from wagtailorderable.models import Orderable
 
-from bc.utils.email import NotifyEmailMessage
-
 from ..utils.blocks import StoryBlock
 from ..utils.constants import RICH_TEXT_FEATURES
+from ..utils.email import NotifyEmailMessage
 from ..utils.models import BasePage
+from .constants import JOB_BOARD_CHOICES
 
 
 class JobSubcategory(models.Model):
@@ -35,7 +36,7 @@ class JobSubcategory(models.Model):
     This corresponds to Job Group in the TalentLink import API
     """
 
-    title = models.CharField(max_length=128)
+    title = models.CharField(max_length=128, unique=True)
 
     def get_categories_list(self):
         if self.categories:
@@ -79,14 +80,18 @@ class JobCategory(Orderable, models.Model):
         )
 
     @staticmethod
-    def get_categories_summary(queryset=None):
+    def get_categories_summary(queryset=None, homepage=None):
         """Returns a QuerySet that returns dictionaries, when used as an iterable.
 
            The dictionary keys are: category (category id), count, title, description
            This is ordered by highest count first.
         """
+        if not homepage:
+            homepage = RecruitmentHomePage.objects.live().first()
         if not queryset:
-            queryset = TalentLinkJob.objects.all()
+            queryset = TalentLinkJob.objects.filter(homepage=homepage).all()
+        else:
+            queryset = queryset.filter(homepage=homepage)
 
         job_categories = (
             queryset.annotate(category=F("subcategory__categories"))
@@ -158,14 +163,17 @@ def callback_jobcategory_autogenerate_slug_if_empty(sender, instance, *args, **k
 
 
 class TalentLinkJob(models.Model):
-
     created = models.DateTimeField(auto_now_add=True)
     last_imported = models.DateTimeField(blank=True)
-    last_modified = models.DateTimeField(auto_now=True)
+    last_modified = models.DateTimeField(
+        auto_now=True
+    )  # For debugging; not modified during normal application flow.
 
     talentlink_id = models.IntegerField(unique=True)
     job_number = models.CharField(max_length=10, blank=False)
-
+    homepage = models.ForeignKey(
+        "RecruitmentHomePage", on_delete=models.CASCADE, blank=True, null=True
+    )
     title = models.CharField(max_length=255, blank=False)
     short_description = models.TextField()
     description = models.TextField()
@@ -180,13 +188,14 @@ class TalentLinkJob(models.Model):
     working_hours = models.CharField(max_length=255)
     closing_date = models.DateField()
     expected_start_date = models.DateField(null=True)
+    interview_date = models.DateField(null=True)
 
     contact_email = models.EmailField()
 
     searchable_salary = models.CharField(
         max_length=255, help_text="Salary group for filtering"
     )
-    searchable_location = models.CharField(max_length=255)
+    location = models.CharField(max_length=255)
     location_postcode = models.CharField(max_length=8, blank=True)
     location_lat = models.DecimalField(
         max_digits=9, decimal_places=6, null=True, blank=True
@@ -201,6 +210,9 @@ class TalentLinkJob(models.Model):
     attachments = models.ManyToManyField(
         "documents.CustomDocument", blank=True, related_name="jobs"
     )
+    logo = models.ForeignKey(
+        "images.CustomImage", null=True, related_name="+", on_delete=models.SET_NULL,
+    )
     application_url_query = models.CharField(max_length=255)
 
     def get_categories_list(self):
@@ -212,10 +224,6 @@ class TalentLinkJob(models.Model):
 
     def __str__(self):
         return f"{self.job_number}: {self.title}"
-
-    @cached_property
-    def homepage(self):
-        return RecruitmentHomePage.objects.live().public().first()
 
     @property
     def url(self):
@@ -229,13 +237,27 @@ class TalentLinkJob(models.Model):
         scheme, netloc, path, query, fragment = urlsplit(base_url)
         return urlunsplit((scheme, netloc, path, self.application_url_query, fragment))
 
+    def full_clean(self, *args, **kwargs):
+        if not self.homepage:
+            self.homepage = RecruitmentHomePage.objects.live().first()
+
+        super().full_clean(*args, **kwargs)
+
 
 @receiver(pre_delete, sender=TalentLinkJob)
-def callback_talentlinkjob_delete_attachments(sender, instance, *args, **kwargs):
+def callback_talentlinkjob_delete_attachments_and_logo(
+    sender, instance, *args, **kwargs
+):
     # if instance.attachments:
     for doc in instance.attachments.all():
         if doc.jobs.all().count() == 1:
             doc.delete()
+
+    # Delete associated logo if it isn't used anywhere else
+    if instance.logo and (
+        TalentLinkJob.objects.filter(logo=instance.logo).count() == 1
+    ):
+        instance.logo.delete()
 
 
 class RecruitmentHomePage(RoutablePageMixin, BasePage):
@@ -244,6 +266,7 @@ class RecruitmentHomePage(RoutablePageMixin, BasePage):
     # Only allow creating HomePages at the root level
     parent_page_types = ["wagtailcore.Page"]
 
+    job_board = models.CharField(max_length=20, blank=True, unique=True)
     hero_title = models.CharField(
         max_length=255, help_text="e.g. Finding a job in Buckinghamshire"
     )
@@ -290,10 +313,16 @@ class RecruitmentHomePage(RoutablePageMixin, BasePage):
         ),
         StreamFieldPanel("body"),
     ]
+    settings_panels = BasePage.settings_panels + [
+        FieldPanel(
+            "job_board",
+            widget=forms.Select(choices=[(s, s) for s in JOB_BOARD_CHOICES],),
+        ),
+    ]
 
     def get_context(self, request, *args, **kwargs):
         context = super().get_context(request, *args, **kwargs)
-        context["job_categories"] = JobCategory.get_categories_summary()
+        context["job_categories"] = JobCategory.get_categories_summary(homepage=self)
 
         return context
 
@@ -350,9 +379,16 @@ class JobAlertSubscription(models.Model):
     search = models.TextField(
         default="{}", editable=False
     )  # stop site admins from entering bad values
+    homepage = models.ForeignKey(
+        "RecruitmentHomePage", on_delete=models.CASCADE, blank=True, null=True
+    )
     confirmed = models.BooleanField(default=False)
     created = models.DateTimeField(auto_now_add=True)
     token = models.CharField(max_length=255, unique=True, editable=False)
+
+    @cached_property
+    def site_url(self):
+        return self.homepage.url.rstrip("/")
 
     @property
     def confirmation_url(self):
@@ -365,6 +401,8 @@ class JobAlertSubscription(models.Model):
     def full_clean(self, *args, **kwargs):
         if not self.token:
             self.token = secrets.token_urlsafe(32)
+        if not self.homepage:
+            self.homepage = RecruitmentHomePage.objects.live().first()
 
         super().full_clean(*args, **kwargs)