+++ bc/static_src/javascript/main.js
@@ -6,9 +6,13 @@ import MobileMenu from './components/mobile-menu';
 import MobileSubMenu from './components/mobile-sub-menu';
 // import CookieWarning from './components/cookie-message';
 import Accordion from './components/accordion';
+import AreaSearchForm from './components/area-search-form';
 import Carousel from './components/carousel';
 import ProgressBar from './components/progress-bar';
 import VideoModal from './components/video-modal';
+import FormSubmit from './components/form-submit';
+
+import 'whatwg-fetch';
 
 import '../sass/main.scss';
 
@@ -33,8 +37,18 @@ document.addEventListener('DOMContentLoaded', () => {
         new Accordion(accordion);
     }
 
+    for (const areasearchform of document.querySelectorAll(
+        AreaSearchForm.selector(),
+    )) {
+        new AreaSearchForm(areasearchform);
+    }
+
     for (const carousel of document.querySelectorAll(Carousel.selector())) {
         new Carousel(carousel);
+    }
+
+    for (const formsubmit of document.querySelectorAll(FormSubmit.selector())) {
+        new FormSubmit(formsubmit);
     }
 
     for (const mobilemenu of document.querySelectorAll(MobileMenu.selector())) {