+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Release History
 
+## Unreleased
+
+- Add tests for job categories import fix
+- Add unit tests for Aptean Respond integration
+- Subheading block (BE)
+- Styling fix for links to local areas
+- Fix spacing in breadcrumb navigation
+- Fix spacing on job detail template
+- Fix for multiple form submissions
+
+## 14.0 (2020-04-22)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/13.0...14.0
+
+- 135 Employer logos (BE)
+- 176 Jobs fix import location
+- 204 Jobs fix salary imports (BE)
+
+## 13.0 (2020-04-20)
+
+Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/12.0...13.0
+
+- Upgrade Wagtail to 2.7.2 and Django to 2.2.12
+- Add documentation about Aptean Respond cases backend, and cookies
+- Make the separate forms on the jobs search page cooperate
+- Styling tweaks on recruitment index page ('benefits' page)
+- Add favicon
+- 140 Job site breadcrumbs on all pages
+- 142 Internal jobs site
+- 155 Jobs extra filters (BE)
+- 161 Update job search filter styles and add custom select form field
+- 173 Search local council area by postcode
+- 179 Jobs additional fields BE
+
 ## 12.0 (2020-04-02)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/11.0...12.0
@@ -8,6 +42,7 @@ Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/11.0...12.0
 - 172 Add live chat support feature
 - 174 Fix button padding
 - 175 Remove hard coded Bucks logo from job detail pages
+- 193 Fix issue with duplicate JobSubcategory titles
 
 ## 11.0 (2020-04-01)
 
@@ -32,7 +67,6 @@ Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/9.0...10.0
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/8.0...9.0
 
 - 129 Jobs postcode search
-- 161 Update job search filter styles and add custom select form field
 - 169 Adjust position of form field help text
 - Forms integration with Aptean Respond API
 - Fix an issue with CI deployments